add single dir option

+++ lib/Any/Template/ProcessDir.pm
@@ -11,23 +11,38 @@ use Try::Tiny;
 use strict;
 use warnings;
 
-has 'dest_dir'             => ( is => 'ro', required => 1 );
+has 'dest_dir'             => ( is => 'ro' );
+has 'dir'                  => ( is => 'ro' );
 has 'dir_create_mode'      => ( is => 'ro', isa => 'Int', default => oct(775) );
 has 'file_create_mode'     => ( is => 'ro', isa => 'Int', default => oct(444) );
 has 'ignore_files'         => ( is => 'ro', isa => 'CodeRef', default => sub { sub { 0 } } );
 has 'process_file'         => ( is => 'ro', isa => 'CodeRef', default => sub { \&_default_process_file } );
 has 'process_text'         => ( is => 'ro', isa => 'CodeRef', default => sub { \&_default_process_text } );
 has 'readme_filename'      => ( is => 'ro', default => 'README' );
-has 'source_dir'           => ( is => 'ro', required => 1 );
+has 'same_dir'             => ( is => 'ro', init_arg => undef );
+has 'source_dir'           => ( is => 'ro' );
 has 'template_file_suffix' => ( is => 'ro', default => '.src' );
+
+sub BUILD {
+    my ( $self, $params ) = @_;
+
+    die "you must pass one of dir and source_dir/dest_dir"
+      if (
+        defined( $self->dir ) ==
+        ( defined( $self->source_dir ) && defined( $self->dest_dir ) ) );
+    $self->{same_dir} = defined( $self->dir );
+}
 
 sub process_dir {
     my ($self) = @_;
 
     my $source_dir = $self->source_dir;
     my $dest_dir   = $self->dest_dir;
-    remove_tree($dest_dir);
-    die "could not remove '$dest_dir'" if -d $dest_dir;
+
+    if ( !$self->same_dir ) {
+        remove_tree($dest_dir);
+        die "could not remove '$dest_dir'" if -d $dest_dir;
+    }
 
     my $ignore_files = $self->ignore_files;
     my @source_files =
@@ -38,8 +53,10 @@ sub process_dir {
         $self->generate_dest_file($source_file);
     }
 
-    $self->generate_readme();
-    try { $self->generate_source_symlink() };
+    if ( !$self->same_dir ) {
+        $self->generate_readme();
+        try { $self->generate_source_symlink() };
+    }
 }
 
 sub generate_dest_file {
@@ -64,11 +81,14 @@ sub generate_dest_file {
         $dest_text = read_file($source_file);
     }
 
-    die "$dest_file already exists!" if -f $dest_file;
-
-    make_path( dirname($dest_file) );
-    chmod( $self->dir_create_mode(), dirname($dest_file) )
-      if defined( $self->dir_create_mode() );
+    if ( $self->same_dir ) {
+        unlink($dest_file);
+    }
+    else {
+        make_path( dirname($dest_file) );
+        chmod( $self->dir_create_mode(), dirname($dest_file) )
+          if defined( $self->dir_create_mode() );
+    }
 
     write_file( $dest_file, $dest_text );
     chmod( $self->file_create_mode(), $dest_file )
@@ -128,8 +148,7 @@ Any::Template::ProcessDir -- Process a directory of templates
     use Any::Template::ProcessDir;
 
     my $pd = Any::Template::ProcessDir->new(
-        source_dir   => '/path/to/source/dir',
-        dest_dir     => '/path/to/dest/dir',
+        source_dir   => '/path/to/dir',
         process_text => sub {
             my $template = Any::Template->new( Backend => '...', String => $_[0] );
             $template->process({ ... });
@@ -137,42 +156,53 @@ Any::Template::ProcessDir -- Process a directory of templates
     );
     $pd->process_dir();
 
-    # or
-
     my $pd = Any::Template::ProcessDir->new(
-        source_dir   => '/path/to/source/dir',
-        dest_dir     => '/path/to/dest/dir',
-        process_text => sub {
+        source_dir => '/path/to/source/dir',
+        dest_dir   => '/path/to/dest/dir',
+        process_file => sub {
             my $file = $_[0];
             # do something with $file, return content
         }
     );
     $pd->process_dir();
 
-    
 =head1 DESCRIPTION
 
-Recursively processes a directory of templates, generating a parallel directory
-of result files. Each file in the source directory may be template-processed,
-copied, or ignored depending on its pathname.
+Recursively processes a directory of templates, generating a set of result
+files in the same directory or in a parallel directory. Each file in the source
+directory may be template-processed, copied, or ignored depending on its
+pathname.
 
 =head1 CONSTRUCTOR
 
-Required parameters:
+=head2 Specifying directory/directories
 
 =over
 
-=item source_dir
+=item *
 
-Directory containing the template files.
+If you want to generate the result files in the B<same> directory as the
+templates, just specify I<dir>.
 
-=item dest_dir
+    my $pd = Any::Template::ProcessDir->new(
+        dir => '/path/to/dir',
+        ...
+    );
 
-Directory where you want to generate result files.
+=item *
+
+If you want to generate the result files in a B<separate> directory from the
+templates, specify I<source_dir> and I<dest_dir>.
+
+    my $pd = Any::Template::ProcessDir->new(
+        source_dir => '/path/to/source/dir',
+        source_dir => '/path/to/dest/dir',
+        ...
+    );
 
 =back
 
-Plus one of these:
+=head2 Specifying how to process templates
 
 =over
 
@@ -189,14 +219,14 @@ the result string. This can use L<Any::Template> or another method altogether.
 
 =back
 
-Optional parameters:
+=head2 Optional parameters
 
 =over
 
 =item dir_create_mode
 
 Permissions mode to use when creating destination directories. Defaults to
-0775.
+0775. No effect if you are using a single directory.
 
 =item file_create_mode
 
@@ -211,7 +241,7 @@ ignored. By default, all files will be considered.
 =item readme_filename
 
 Name of a README file to generate in the destination directory - defaults to
-"README".
+"README". No effect if you are using a single directory.
 
 =item template_file_suffix
 
@@ -229,8 +259,9 @@ L</ignore_file_suffix>) will simply be copied to the destination.
 
 =item process_dir
 
-Process the directory. The destination directory will be removed completely and
-recreated, to eliminate any old files from previous processing.
+Process the directory. If using multiple directories, the destination directory
+will be removed completely and recreated, to eliminate any old files from
+previous processing.
 
 =back
